Use multiprocessing instead of threading
