+++ package/info.log
@@ -1,0 +1,2 @@
+2019-10-22 11:27:56,590 - root - ERROR - 'int' object is not callable
+2019-10-22 11:29:56,665 - root - ERROR - 'int' object is not callable
